# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminLoginRequest {
  email: string;
  password: string;
}

export interface AdminUser {
  id: number;
  email: string;
  role: string;
}

export interface AdminSessionResponse {
  user: AdminUser;
}

export interface ContentRecord { [key: string]: unknown }

export interface ContentListResponse {
  items: ContentRecord[];
}

export type MediaUploadRequestContentType = typeof MediaUploadRequestContentType[keyof typeof MediaUploadRequestContentType];


export const MediaUploadRequestContentType = {
  'image/jpeg': 'image/jpeg',
  'image/png': 'image/png',
  'image/webp': 'image/webp',
  'image/gif': 'image/gif',
} as const;

export interface MediaUploadRequest {
  name: string;
  size: number;
  contentType: MediaUploadRequestContentType;
}

export interface MediaUploadResponse {
  uploadURL: string;
  objectPath: string;
}

export type MediaUploadVerificationRequestContentType = typeof MediaUploadVerificationRequestContentType[keyof typeof MediaUploadVerificationRequestContentType];


export const MediaUploadVerificationRequestContentType = {
  'image/jpeg': 'image/jpeg',
  'image/png': 'image/png',
  'image/webp': 'image/webp',
  'image/gif': 'image/gif',
} as const;

export interface MediaUploadVerificationRequest {
  objectPath: string;
  size: number;
  contentType: MediaUploadVerificationRequestContentType;
}