# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary Start an admin session
 */
export const AdminLoginBody = zod.object({
  "email": zod.string().email(),
  "password": zod.string()
})

export const AdminLoginResponse = zod.object({
  "user": zod.object({
  "id": zod.number().int(),
  "email": zod.string().email(),
  "role": zod.string()
})
})


/**
 * @summary Get the current admin session
 */
export const AdminSessionResponse = zod.object({
  "user": zod.object({
  "id": zod.number().int(),
  "email": zod.string().email(),
  "role": zod.string()
})
})


/**
 * @summary End the current admin session
 */
export const AdminLogoutResponse = zod.unknown()


/**
 * @summary List protected content records
 */
export const ListAdminContentParams = zod.object({
  "resource": zod.enum(['people', 'projects', 'news', 'researchAreas', 'teaching', 'profile', 'media'])
})

export const ListAdminContentResponse = zod.object({
  "items": zod.array(zod.record(zod.string(), zod.unknown()))
})


/**
 * @summary Create a protected content record
 */
export const CreateAdminContentParams = zod.object({
  "resource": zod.enum(['people', 'projects', 'news', 'researchAreas', 'teaching', 'profile', 'media'])
})

export const CreateAdminContentBody = zod.record(zod.string(), zod.unknown())

export const CreateAdminContentResponse = zod.record(zod.string(), zod.unknown())


/**
 * @summary Update a protected content record
 */
export const UpdateAdminContentParams = zod.object({
  "resource": zod.enum(['people', 'projects', 'news', 'researchAreas', 'teaching', 'profile', 'media']),
  "id": zod.coerce.number().int()
})

export const UpdateAdminContentBody = zod.record(zod.string(), zod.unknown())

export const UpdateAdminContentResponse = zod.record(zod.string(), zod.unknown())


/**
 * @summary Archive a protected content record
 */
export const ArchiveAdminContentParams = zod.object({
  "resource": zod.enum(['people', 'projects', 'news', 'researchAreas', 'teaching', 'profile', 'media']),
  "id": zod.coerce.number().int()
})

export const ArchiveAdminContentResponse = zod.unknown()


/**
 * @summary Request a presigned image upload URL
 */
export const RequestAdminMediaUploadUrlBody = zod.object({
  "name": zod.string(),
  "size": zod.number().int(),
  "contentType": zod.enum(['image/jpeg', 'image/png', 'image/webp', 'image/gif'])
})

export const RequestAdminMediaUploadUrlResponse = zod.object({
  "uploadURL": zod.string().url(),
  "objectPath": zod.string()
})


/**
 * @summary Verify an uploaded image before creating its media record
 */
export const VerifyAdminMediaUploadBody = zod.object({
  "objectPath": zod.string(),
  "size": zod.number().int(),
  "contentType": zod.enum(['image/jpeg', 'image/png', 'image/webp', 'image/gif'])
})

export const VerifyAdminMediaUploadResponse = zod.unknown()