# lib/api-zod/src/generated/types/adminSessionResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUser } from './adminUser';

export interface AdminSessionResponse {
  user: AdminUser;
}