# lib/api-zod/src/generated/types/contentListResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContentRecord } from './contentRecord';

export interface ContentListResponse {
  items: ContentRecord[];
}