# lib/api-zod/src/generated/types/mediaUploadRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MediaUploadRequestContentType } from './mediaUploadRequestContentType';

export interface MediaUploadRequest {
  name: string;
  size: number;
  contentType: MediaUploadRequestContentType;
}