# lib/api-zod/src/generated/types/mediaUploadVerificationRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MediaUploadVerificationRequestContentType } from './mediaUploadVerificationRequestContentType';

export interface MediaUploadVerificationRequest {
  objectPath: string;
  size: number;
  contentType: MediaUploadVerificationRequestContentType;
}